+++ components/common/ConfirmPurchaseModal.vue
@@ -95,7 +95,7 @@ const totalNFTsPrice = computed(() =>
   sum(items.value.map((nft) => Number(nft.price)))
 )
 const totalRoyalties = computed(() =>
-  sum(items.value.map((nft) => Number(nft.royalty?.amount || 0)))
+  sum(items.value.map((nft) => Number(nft.royalty?.amount ?? 0)))
 )
 const disabled = ref(false)
