+++ utils/ipfs.ts
@@ -179,14 +179,3 @@ export type EntityWithMeta = {
   metadata: string
   meta?: NFTMetadata
 }
-
-export const parseNftAvatar = async (
-  entity: EntityWithMeta
-): Promise<string> => {
-  if (entity.meta?.image) {
-    return sanitizeIpfsUrl(entity.meta.image)
-  } else {
-    const meta = (await processSingleMetadata(entity.metadata)) as NFTMetadata
-    return sanitizeIpfsUrl(meta?.image)
-  }
-}
